requirements.txt
# PQ Agile Wrapper — Dependencies
# Core
pyyaml>=6.0
cryptography>=42.0.0

# Post-Quantum Cryptography
# liboqs-python requires liboqs C library to be installed
# See Dockerfile for system-level installation
# For local dev: https://github.com/open-quantum-safe/liboqs-python
oqs>=0.9.0; platform_system != "Windows"

# Async networking
asyncio-dgram>=2.1.0

# Monitoring & Logging
rich>=13.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
